+++ Engine/Source/Scene/Scene.cpp
@@ -1,6 +1,8 @@
 #include "Scene.h"
 #include "Core/Logger.h"
 
+#include <algorithm>
+
 namespace Core
 {
     Scene::Scene()
@@ -80,5 +82,68 @@ namespace Core
                 return a;
 
         return nullptr;
+    }
+
+    Actor *Scene::GetActorInHierarchy(const UUID &uuid)
+    {
+        Actor *a = GetActorByUUID(uuid);
+
+        if (!a)
+        {
+            for (auto actor : actors)
+            {
+                if (a)
+                    break;
+                a = actor->FindChildInHierarchyByUUID(uuid);
+            }
+        }
+
+        return a;
+    }
+
+    void Scene::RemoveActorByUUID(const UUID &uid)
+    {
+        Actor *a;
+
+        int index = -1;
+        for (auto actor : actors)
+        {
+            index++;
+            if (actor->id == uid)
+            {
+                actors.erase(actors.begin() + index);
+                break;
+            }
+        }
+    }
+
+    void Scene::MoveChildInHierarchy(const UUID &uid, int newIndex)
+    {
+
+        Actor *actorToMove = GetActorByUUID(uid);
+
+        if (!actorToMove)
+        {
+            return;
+        }
+
+        if (newIndex < 0 || newIndex >= actors.size())
+        {
+            return;
+        }
+
+        // Find the current index of the actor
+        auto actorIterator = std::find(actors.begin(), actors.end(), actorToMove);
+
+        if (actorIterator != actors.end())
+        {
+            size_t currentIndex = std::distance(actors.begin(), actorIterator);
+
+            // Remove the actor from the current position
+            actors.erase(actorIterator);
+
+            // Insert the actor at the new index
+            actors.insert(actors.begin() + newIndex, actorToMove);
+        }
     }
 }